requirements.txt
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==4.5.4
coveralls==1.8.2
docopt==0.6.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.23.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1 
mccabe==0.6.1
more-itertools==7.2.0
nose==1.3.7
packaging==19.2
pep8==1.7.1
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pytest==5.1.3
python-decouple==3.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.5
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
